# core/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Crawlab API
  version: 0.6.0
  description: Crawlab API
  license:
    name: BSD-3-Clause
    url: https://github.com/crawlab-team/crawlab-pro/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Local API
  - url: http://localhost:8080/api
    description: Docker API
  - url: https://demo-pro.crawlab.cn/api
    description: Demo API
...
paths:
...
  /tasks/run:
    post:
      tags: [ task ]
      operationId: postTaskRun
      summary: Run task
      security:
        - apiToken: [ ]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTaskRunRequest'
      responses:
        200:
          description: Run task successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
...
  /tasks/{id}/restart:
    post:
      tags: [ task ]
      operationId: postTaskRestart
      summary: Restart task
      security:
        - apiToken: [ ]
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restart task successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
  /tasks/{id}/cancel:
    post:
      tags: [ task ]
      operationId: postTaskCancel
      summary: Cancel task
      security:
        - apiToken: [ ]
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
      responses:
        200:
          description: Cancel task successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
  /tasks/{id}/logs:
    get:
      tags: [ task ]
      operationId: getTaskLogs
      summary: Get task logs
      security:
        - apiToken: [ ]
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
        - name: size
          in: query
          description: Page size number
          required: false
          schema:
            type: integer
      responses:
        200:
          description: Get task logs successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTaskLogsResponse'
        404:
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
  /tasks/{id}/data:
    get:
      tags: [ task ]
      operationId: getTaskData
      summary: Get task data
      security:
        - apiToken: [ ]
      parameters:
        - name: id
          in: path
          description: Task ID
          required: true
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
        - name: size
          in: query
          description: Page size number
          required: false
          schema:
            type: integer
      responses:
        200:
          description: Get task data successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTaskDataResponse'
        404:
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'

  # user
...
  /users/{id}/change-password:
    post:
      tags: [ user ]
      operationId: changePassword
      summary: Change password
      security:
        - apiToken: [ ]
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUserChangePasswordRequest'
      responses:
        200:
          description: Change password successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
...
components:
  schemas:
...
    PostTaskRunRequest:
      description: Post task run request body
      properties:
        spider_id:
          type: string
          description: Task spider ID
        cmd:
          type: string
          description: Task command
        params:
          type: string
          description: Task params
        mode:
          $ref: '#/components/schemas/TaskMode'
        priority:
          type: integer
          description: Task priority
...
    PostUserChangePasswordRequest:
      description: Post user change password request body
      properties:
        password:
          type: string
          description: New password
...
    BaseResponse:
      type: object
      properties:
        status:
          type: string
          description: Status of the response
        message:
          type: string
          description: Message of the response
    EmptyResponse:
      description: Empty response body
      $ref: '#/components/schemas/BaseResponse'
      example:
        status: ok
        message: success
...
    ListResponse:
      description: List response body
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
            total:
              type: integer
    ErrorResponse:
      description: Error response body
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: string

    # error response
    NotFoundErrorResponse:
      description: Not found error response body
      $ref: '#/components/schemas/ErrorResponse'
      example:
        status: error
        message: not found
    BadRequestErrorResponse:
      description: Bad request error response
      $ref: '#/components/schemas/ErrorResponse'
      example:
        status: error
        message: bad request
...
    GetTaskLogsResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ListResponse'
        - properties:
          data:
            type: array
            description: Task logs
            items:
              type: string
    GetTaskDataResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ListResponse'
        - properties:
            data:
              type: array
              items:
                type: object
...
    TaskMode:
      type: string
      enum:
        - random
        - all-nodes
        - selected-nodes
...
  securitySchemes:
    apiToken:
      type: apiKey
      in: header
      name: Authorization
      description: API Token
      x-displayName: API Token